// frontend/src/api/model/partnerLogo.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Open-Q API
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerLogoUrl } from './partnerLogoUrl';

/**
 * Schema for a partner institution logo.
 */
export interface PartnerLogo {
    id: string;
    name: string;
    logo_url: string;
    url?: PartnerLogoUrl;
}